get rid of `fieldMaps` because I don't like it
