+++ apps/spruce/src/components/DayPicker/index.tsx
@@ -3,7 +3,7 @@ import styled from "@emotion/styled";
 import { palette } from "@leafygreen-ui/palette";
 import { transitionDuration } from "@leafygreen-ui/tokens";
 import { size } from "@evg-ui/lib/constants/tokens";
-import { days } from "constants/fieldMaps";
+import { days } from "constants/time";
 
 const { gray, white } = palette;
