+++ apps/spruce/src/pages/waterfall/DateFilter/index.tsx
@@ -34,7 +34,7 @@ export const DateFilter = () => {
   return (
     <DatePicker
       data-cy="date-picker"
-      // Use an uncontrolled component so that the transition does not DatePicker rendering.
+      // Use an uncontrolled component so that the transition does not block DatePicker rendering.
       initialValue={date.length ? new Date(date) : undefined}
       label="Go to Date"
       max={new Date()}
